pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Ayron Felipe <ayron41@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.6"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.9"
markdown = "^3.5"
django-filter = "^23.3"
boto3 = "^1.28.73"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.5.2"
isort = "^5.12.0"
black = "^23.10.1"
mypy = "^1.6.1"
flake8 = "^6.1.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
force_single_line = true
ensure_newline_before_comments = true
line_length = 120
skip = [".gitignore", ".dockerignore", "migrations"]
use_parentheses = true
multi_line_output = 3
include_trailing_comma = true
lines_between_types = 1

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | migrations
  | dist
)/
'''

force-exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | migrations
  | dist
)/
'''
